Added ability to start rig between two meshes
Previously the rig had to start at the center of a mesh object, creating an armature with a single bone that could rotate the entire model. Now starts the rig between two mesh objects so that each bone branches outward from the center of where the two mesh objects meet and the armature does not have an arbitrary bone with the capability to rotate the model.